Rework RSVP as per Phil's feedback
You can link to /webjam8/rsvp
Register/login after hitting "RSVP" takes you to /webjam8/rsvp
The event action buttons have been simplified

+++ app/controllers/rsvps_controller.rb
@@ -1,7 +1,14 @@
 class RsvpsController < ApplicationController
   before_filter :login_required
 
-  def create
+  def show
+    @event = Event.find_by_tag(params[:event_id])
+    raise NotFound unless @event
+    @rsvp = current_user.rsvp_for(@event)
+    render :action => (@rsvp ? :show : :new)
+  end
+  
+  def update
     @event = Event.find_by_tag(params[:event_id])
     raise NotFound unless @event
     if @event.full?
@@ -17,8 +24,9 @@ class RsvpsController < ApplicationController
   def destroy
     event = Event.find_by_tag(params[:event_id])
     raise NotFound unless event
-    rsvp = current_user.rsvps.find(params[:id])
-    rsvp.destroy
+    if rsvp = current_user.rsvp_for(event)
+      rsvp.destroy
+    end
     redirect_to event_path(event)
   end
 end
